refactor(with-data) Move retrievers out of main file
